THE BIG UPDATE - PROTOTYPE 1
As of today, we have an (almost) fully integrated system! The webhooks for the python script are coded in (mostly, barring a single thing for test), the speech to text and voice recognition modules are in place, and major overhauls now seem unlikely on the current codebase. Refinements will continue from here.